THE FINAL SIEGE COMMIT
Updated README. No GIF sadly, I am burnt tf out lol

Thanks for making this happen, Olive. This was the best YSWS I ever participated in. Thanks to HC to make this possible I will forever be grateful for providing me with materials when I was in pinch, when I had nothing but DETERMINATION